fix: move resend email endpoint to the unprotected controller and add mutual friends endpoint
